# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSI",
        "version": "1.0"
    },
    "paths": {
...
        "/api/etat/{id}": {
            "get": {
                "tags": [
                    "Etat"
                ],
                "description": "Cette route permet de retourner l'état dont l'id est passé en paramètre au format Json.",
                "operationId": "d4e49678b8bf45466cd07f14d14849a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la l'état que nous voulons récupéré !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne l'état dont l'id est passé en paramètre au format JSON.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Etat"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/create-etat": {
            "get": {
                "tags": [
                    "Etat"
                ],
                "description": "Cette route permet de récupéré l'état d'un ticket dont l'id est passé en paramètre.",
                "operationId": "8747952508761adac01d421f76509b78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id du ticket dont nous voulons récupéré l'état",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne au format Json l'état qui vient d'être créé et enregistrer en base de données.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Etat"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Etat"
                ],
                "description": "Cette route permet de créer un état en base de donnée.",
                "operationId": "7d5362b88d3c05c4255d4197009aeac1",
                "responses": {
                    "200": {
                        "description": "Retourne au format Json l'état qui vient d'être créé et enregistrer en base de données.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Etat"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/edit-etat/{id}": {
            "put": {
                "tags": [
                    "Etat"
                ],
                "description": "Cette route permet de modifier un état déjà existant en base de donnée.",
                "operationId": "319497f57fd086b26580c5167862cf29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la l'état que nous voulons modifier !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne au format Json l'état qui vient d'être modifié et enregistrer en base de données.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Etat"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete-etat/{id}": {
            "delete": {
                "tags": [
                    "Etat"
                ],
                "description": "Cette route permet de supprimer de la base de donnée un état existant ",
                "operationId": "c20665becced619e542d4136cef6d210",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de la l'état que nous voulons supprimer !",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un message de succès si la suppression de l'état a eu lieu en base de données."
                    },
                    "500": {
                        "description": "Retourne un message d'erreur si la suppression de l'état n'a pas eu lieu en base de données."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Etat": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string"
                    },
                    "ordre": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}